backend/requirements.txt
flask
firebase_admin
google-cloud-secret-manager
gunicorn